Trabajo en la mañana

+++ ProyectoBDI - SisVent/ProyectoBDI - SisVent/data/Producto.cs
@@ -11,7 +11,7 @@ namespace ProyectoBDI___SisVent.data
         public int ID { get; set; }
 	    public int CategoriaID { get; set; }
         public string Nombre { get; set; }
-        public float PrecioVenta { get; set; }
+        public decimal PrecioVenta { get; set; }
         public int Stock { get; set; }
         public string Descripcion { get; set; }
         public int Estado { get; set; }
